+++ tests/testthat.R
@@ -2,5 +2,4 @@ library(testthat)
 library(assignR)
 library(raster)
 
-
 test_check("assignR")